#1999 (インストーラのDBクリーンアップ時に Warning が出る)
 * #1664 の追加修正
 * r22347 をマージ

+++ html/install/index.php
@@ -202,7 +202,9 @@ switch ($mode) {
         while ($file_name = readdir($res_dir)){
             //dummy以外は削除
             if ($file_name != 'dummy'){
-                unlink($cache_dir . $file_name);
+                if (is_file($cache_dir . $file_name)) {
+                    unlink($cache_dir . $file_name);
+                }
             }
         }
         closedir($res_dir);
